requirements.txt
numpy>=1.21.0
pyparsing>=2.4.7
virtualenv>=16.7.8
tqdm
pyjosa
jupyter

python-dotenv
openai

torch>=2.0.0
transformers>=4.54.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99